Use float for skill values for a greater flexibility

+++ src/Api/SkillFactoryInterface.php
@@ -7,5 +7,5 @@ use App\Domain\Skill\AbstractSkill;
 
 interface SkillFactoryInterface
 {
-    public function create(string $skillIdentifier, string $type, int $chance): AbstractSkill;
+    public function create(string $skillIdentifier, string $type, float $chance): AbstractSkill;
 }
