+++ src/Domain/Skill/SkillFactory.php
@@ -18,7 +18,7 @@ class SkillFactory implements SkillFactoryInterface
      * @throws UnknownSkillException
      * @throws InvalidSkillValuesException
      */
-    public function create(string $skillIdentifier, string $type, int $chance): AbstractSkill
+    public function create(string $skillIdentifier, string $type, float $chance): AbstractSkill
     {
         if(!isset(self::SKILLS[$skillIdentifier])) {
             throw new UnknownSkillException("There is no such skill with given skill identifier '{$skillIdentifier}'");
